[#144] refactor: type폴더 구조 설정
