+++ src/api/Meeting/GetourTeam.ts
@@ -1,0 +1,1 @@
+import { privateAxios } from '../axiosConfig';
